Initial fork fixup
* rename packages
* fix build jdk version
* fix arg parse
* create data
* fix schema/writing somewhat
* fix config settings, printouts etc.

Works with scylla iff fixes #7410 and #7424 are in.
