+++ .gitignore
@@ -2,3 +2,7 @@ target/
 out/
 .gradle/
 /build/
+/bin/
+.classpath
+.project
+.settings/
