+++ src/main/java/com/scylladb/alternator/StreamsAdapterDemoHelper.java
@@ -12,11 +12,12 @@
  * specific language governing permissions and limitations under the License.
 */
 
-
-package com.scylladb.java;
+package com.scylladb.alternator;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
@@ -24,21 +25,25 @@ import com.amazonaws.services.dynamodbv2.model.AttributeAction;
 import com.amazonaws.services.dynamodbv2.model.AttributeDefinition;
 import com.amazonaws.services.dynamodbv2.model.AttributeValue;
 import com.amazonaws.services.dynamodbv2.model.AttributeValueUpdate;
+import com.amazonaws.services.dynamodbv2.model.BatchWriteItemRequest;
 import com.amazonaws.services.dynamodbv2.model.CreateTableRequest;
 import com.amazonaws.services.dynamodbv2.model.CreateTableResult;
 import com.amazonaws.services.dynamodbv2.model.DeleteItemRequest;
+import com.amazonaws.services.dynamodbv2.model.DeleteRequest;
 import com.amazonaws.services.dynamodbv2.model.DescribeTableRequest;
 import com.amazonaws.services.dynamodbv2.model.DescribeTableResult;
 import com.amazonaws.services.dynamodbv2.model.KeySchemaElement;
 import com.amazonaws.services.dynamodbv2.model.KeyType;
 import com.amazonaws.services.dynamodbv2.model.ProvisionedThroughput;
 import com.amazonaws.services.dynamodbv2.model.PutItemRequest;
+import com.amazonaws.services.dynamodbv2.model.PutRequest;
 import com.amazonaws.services.dynamodbv2.model.ResourceInUseException;
 import com.amazonaws.services.dynamodbv2.model.ScanRequest;
 import com.amazonaws.services.dynamodbv2.model.ScanResult;
 import com.amazonaws.services.dynamodbv2.model.StreamSpecification;
 import com.amazonaws.services.dynamodbv2.model.StreamViewType;
 import com.amazonaws.services.dynamodbv2.model.UpdateItemRequest;
+import com.amazonaws.services.dynamodbv2.model.WriteRequest;
 
 public class StreamsAdapterDemoHelper {
 
@@ -51,24 +56,23 @@ public class StreamsAdapterDemoHelper {
 
         java.util.List<KeySchemaElement> keySchema = new ArrayList<KeySchemaElement>();
         keySchema.add(new KeySchemaElement().withAttributeName("p").withKeyType(KeyType.HASH)); // Partition
-                                                                                                 // key
+                                                                                                // key
 
         ProvisionedThroughput provisionedThroughput = new ProvisionedThroughput().withReadCapacityUnits(2L)
-            .withWriteCapacityUnits(2L);
+                .withWriteCapacityUnits(2L);
 
         StreamSpecification streamSpecification = new StreamSpecification();
         streamSpecification.setStreamEnabled(true);
         streamSpecification.setStreamViewType(StreamViewType.NEW_IMAGE);
         CreateTableRequest createTableRequest = new CreateTableRequest().withTableName(tableName)
-            .withAttributeDefinitions(attributeDefinitions).withKeySchema(keySchema)
-            .withProvisionedThroughput(provisionedThroughput).withStreamSpecification(streamSpecification);
+                .withAttributeDefinitions(attributeDefinitions).withKeySchema(keySchema)
+                .withProvisionedThroughput(provisionedThroughput).withStreamSpecification(streamSpecification);
 
         try {
             System.out.println("Creating table " + tableName);
             CreateTableResult result = client.createTable(createTableRequest);
             return result.getTableDescription().getLatestStreamArn();
-        }
-        catch (ResourceInUseException e) {
+        } catch (ResourceInUseException e) {
             System.out.println("Table already exists.");
             return describeTable(client, tableName).getTable().getLatestStreamArn();
         }
@@ -83,18 +87,55 @@ public class StreamsAdapterDemoHelper {
     }
 
     public static void putItem(AmazonDynamoDB dynamoDBClient, String tableName, String id, String val) {
-        java.util.Map<String, AttributeValue> item = new HashMap<String, AttributeValue>();
-        item.put("p", new AttributeValue().withS(id));
-        item.put("attribute-1", new AttributeValue().withS(val));
-
-        PutItemRequest putItemRequest = new PutItemRequest().withTableName(tableName).withItem(item);
-        dynamoDBClient.putItem(putItemRequest);
+        dynamoDBClient.putItem(putItem(tableName, items(id, val)));
     }
 
-    public static void putItem(AmazonDynamoDB dynamoDBClient, String tableName,
-        java.util.Map<String, AttributeValue> items) {
-        PutItemRequest putItemRequest = new PutItemRequest().withTableName(tableName).withItem(items);
-        dynamoDBClient.putItem(putItemRequest);
+    public static void putItem(AmazonDynamoDB client, String tableName, int recordNo) {
+        client.putItem(putItem(tableName, items(recordNo)));
+    }
+
+    public static Map<String, AttributeValue> items(String id, String val) {
+        java.util.Map<String, AttributeValue> items = new HashMap<String, AttributeValue>();
+        items.put("p", new AttributeValue().withS(id));
+        items.put("attribute-1", new AttributeValue().withS(val));
+        return items;
+    }
+
+    public static Map<String, AttributeValue> items(int recordNo) {
+        return items(String.valueOf(recordNo), "le gris " + recordNo);
+    }
+
+    public static void putItems(AmazonDynamoDB client, String tableName, int from, int to) {
+        while (from < to) {
+            List<WriteRequest> items = new ArrayList<>(100);
+
+            for (int e = Math.min(to, from + 100); from < e; ++from) {
+                putItem(items, items(from));
+            }
+            batchWrite(client, tableName, items);
+        }
+    }
+
+    public static void batchWrite(AmazonDynamoDB client, String tableName, List<WriteRequest> items) {
+        BatchWriteItemRequest r = new BatchWriteItemRequest()
+                .withRequestItems(Collections.singletonMap(tableName, items));
+        client.batchWriteItem(r);
+    }
+
+    public static void putItem(List<WriteRequest> dst, Map<String, AttributeValue> items) {
+        dst.add(new WriteRequest(new PutRequest(items)));
+    }
+
+    public static void deleteItem(List<WriteRequest> dst, Map<String, AttributeValue> keys) {
+        dst.add(new WriteRequest(new DeleteRequest(keys)));
+    }
+
+    public static void putItem(AmazonDynamoDB dynamoDBClient, String tableName, Map<String, AttributeValue> items) {
+        dynamoDBClient.putItem(putItem(tableName, items));
+    }
+
+    public static PutItemRequest putItem(String tableName, Map<String, AttributeValue> items) {
+        return new PutItemRequest().withTableName(tableName).withItem(items);
     }
 
     public static void updateItem(AmazonDynamoDB dynamoDBClient, String tableName, String id, String val) {
@@ -103,21 +144,28 @@ public class StreamsAdapterDemoHelper {
 
         Map<String, AttributeValueUpdate> attributeUpdates = new HashMap<String, AttributeValueUpdate>();
         AttributeValueUpdate update = new AttributeValueUpdate().withAction(AttributeAction.PUT)
-            .withValue(new AttributeValue().withS(val));
+                .withValue(new AttributeValue().withS(val));
         attributeUpdates.put("attribute-2", update);
 
         UpdateItemRequest updateItemRequest = new UpdateItemRequest().withTableName(tableName).withKey(key)
-            .withAttributeUpdates(attributeUpdates);
+                .withAttributeUpdates(attributeUpdates);
         dynamoDBClient.updateItem(updateItemRequest);
     }
 
-    public static void deleteItem(AmazonDynamoDB dynamoDBClient, String tableName, String id) {
+    public static void deleteItem(AmazonDynamoDB dynamoDBClient, String tableName, Map<String, AttributeValue> atts) {
+        deleteItem(dynamoDBClient, tableName, atts.get("p"));
+    }
+
+    public static void deleteItem(AmazonDynamoDB dynamoDBClient, String tableName, AttributeValue keyValue) {
         java.util.Map<String, AttributeValue> key = new HashMap<String, AttributeValue>();
-        key.put("p", new AttributeValue().withS(id));
+        key.put("p", keyValue);
 
         DeleteItemRequest deleteItemRequest = new DeleteItemRequest().withTableName(tableName).withKey(key);
         dynamoDBClient.deleteItem(deleteItemRequest);
     }
 
-}
+    public static void deleteItem(AmazonDynamoDB dynamoDBClient, String tableName, String id) {
+        deleteItem(dynamoDBClient, tableName, new AttributeValue().withS(id));
+    }
 
+}
